+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://raw.githubusercontent.com/Synogun/FarmRPGPlus/refs/heads/develop/dist/FarmRPGPlus.user.js)
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## [v1.8.2](https://raw.githubusercontent.com/Synogun/FarmRPGPlus/refs/tags/v1.8.2/dist/FarmRPGPlus.user.js)
+
+### Added
 - `HomePage` - Adds Highlight Row feature, which highlights specific rows if they are ready for interaction.
 - `HomePage` - Adds Reset Checklist feature, which adds a list on the bottom of the home page with all the common daily tasks.
 - `HomePage` - Adds Announcements feature, which displays new updates announcements on the top of home page.